+++ weaver/utils.py
@@ -18,7 +18,7 @@ import six
 from celery import Celery
 from lxml import etree
 from pyramid.config import Configurator
-from pyramid.httpexceptions import HTTPError as PyramidHTTPError, HTTPGatewayTimeout
+from pyramid.httpexceptions import HTTPError as PyramidHTTPError, HTTPGatewayTimeout, HTTPTooManyRequests
 from pyramid.registry import Registry
 from pyramid.request import Request
 from pyramid.settings import asbool, aslist
@@ -560,6 +560,7 @@ def request_extra(method,                       # type: AnyStr
                   retries=None,                 # type: Optional[int]
                   backoff=None,                 # type: Optional[Number]
                   intervals=None,               # type: Optional[List[Number]]
+                  retry_after=True,             # type: bool
                   allowed_codes=None,           # type: Optional[List[int]]
                   only_server_errors=True,      # type: bool
                   ssl_verify=None,              # type: Optional[bool]
@@ -573,7 +574,6 @@ def request_extra(method,                       # type: AnyStr
     ---------------
 
     Implements request retry if the previous request failed, up to the specified number of retries.
-
     Using :paramref:`backoff` factor, you can control the interval between request attempts such as::
 
         delay = backoff * (2 ^ retry)
@@ -582,10 +582,22 @@ def request_extra(method,                       # type: AnyStr
     wait between each request attempt. In this case, :paramref:`backoff` is ignored and :paramref:`retries` is
     overridden accordingly with the number of items specified in the list.
 
+    Furthermore, :paramref:`retry_after` (default: ``True``) indicates if HTTP status code ``429 (Too Many Requests)``
+    should be automatically handled during retries. If enabled and provided in the previously failed request response
+    through the ``Retry-After`` header, the next request attempt will be executed only after the server-specified delay
+    instead of following the calculated delay from :paramref:`retries` and :paramref:`backoff`, or from corresponding
+    index of :paramref:`interval`, accordingly to specified parameters. This will avoid uselessly calling the server and
+    automatically receive a denied response. You can disable this feature by passing ``False``, which will result into
+    requests being retried blindly without consideration of the called server instruction.
+
     Because different request implementations use different parameter naming conventions, all following keywords are
     looked for:
         - Both variants of ``backoff`` and ``backoff_factor`` are accepted.
         - All variants of ``retires``, ``retry`` and ``max_retries`` are accepted.
+
+    .. note::
+        Total amount of executed request attempts will be +1 the number of :paramref:`retries` or :paramref:`intervals`
+        items as first request is done immediately, and following attempts are done with the appropriate delay.
 
     File Transport Scheme
     ---------------------
@@ -620,11 +632,9 @@ def request_extra(method,                       # type: AnyStr
 
     :param method: HTTP method to set request.
     :param url: URL of the request to execute.
-    :param retries: number of retries to attempt.
-    :param backoff: factor by which to multiply delays between retries.
-    :param intervals:
-        Explicit intervals in seconds between retries.
-        (note: amount of request attempts will be +1 the number of interval items as first request is done immediately)
+    :param retries: Number of request retries to attempt if first attempt failed (according to allowed codes or error).
+    :param backoff: Factor by which to multiply delays between retries.
+    :param intervals: Explicit intervals in seconds between retries.
     :param allowed_codes: HTTP status codes that are considered valid to stop retrying (default: any non-4xx/5xx code).
     :param ssl_verify: Explicit parameter to disable SSL verification (overrides any settings, default: True).
     :param settings: Additional settings from which to retrieve configuration details for requests.
@@ -647,8 +657,11 @@ def request_extra(method,                       # type: AnyStr
     intervals = intervals or kw_intervals
     if intervals and len(intervals) and all(isinstance(i, (int, float)) for i in intervals):
         intervals = [0] + intervals
-        retries = len(intervals)
+        retries = list(range(len(intervals)))
         backoff = 0  # disable first part of delay calculation
+    else:
+        retries = [0] + list(range(retries))
+    no_retries = len(retries) == 1
     # SSL verification settings
     # ON by default, disable accordingly with any variant if matched
     kw_ssl_verify = get_ssl_verify_option(method, url, settings, request_options=request_options)
@@ -657,9 +670,16 @@ def request_extra(method,                       # type: AnyStr
     # process request
     resp = None
     failures = []
-    for retry in range(retries):
+    for retry in retries:
         if retry:
-            delay = (backoff * (2 ** (retry + 1))) or intervals[retry]
+            delay = 0
+            if retry_after and resp and resp.status_code in [HTTPTooManyRequests.code]:
+                after = resp.headers.get("Retry-After", "")
+                delay = int(after) if str(after).isdigit() else 0
+                LOGGER.debug("Received header [Retry-After=%ss] for [%s %s]", after, method, url)
+            if not delay:
+                delay = (backoff * (2 ** (retry + 1))) or intervals[retry]
+            LOGGER.debug("Retrying failed request after delay=%s for [%s %s]", delay, method, url)
             time.sleep(delay)
         try:
             with requests.Session() as request_session:
@@ -677,14 +697,16 @@ def request_extra(method,                       # type: AnyStr
         # function called without retries raises original error
         # as if calling requests module directly
         except requests.ConnectionError as exc:
-            if not retries:
+            if no_retries:
                 raise
             failures.append(type(exc).__name__)
     # also pass-through here if no retries
-    if not retries and resp:
+    if no_retries and resp:
         return resp
     detail = "Request ran out of retries. Attempts generated following errors: {}".format(failures)
     err = HTTPGatewayTimeout(detail=detail)
+    # make 'raise_for_status' method available for convenience
+    setattr(err, "url", url)
     setattr(err, "reason", err.explanation)
     setattr(err, "raise_for_status", lambda: Response.raise_for_status(err))  # noqa
     return err